requirements.txt
openai==1.77.0
playwright==1.52.0
python-dotenv==1.1.0
Requests==2.32.3
